+++ debug_simple.chest
@@ -1,0 +1,4 @@
+building Debug
+  office Main
+    employee Start
+      show 5 > 3
